apps/inference/pyproject.toml
[project]
name = "pantrific-inference"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "ultralytics>=8.3.0",
    "httpx>=0.28.0",
    "opencv-python>=4.10.0",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.9.0",
]
